refactor: make testable
